extract Type_Colors
